re-organized the test files
